+++ example/lib/step_one.dart
@@ -55,24 +55,24 @@ class _StepOneState extends State<StepOne> {
               ),
             ),
 
-            InkWell(
-              onTap: (){
-                setState(() {
-
-                });
-               // widget.controller.next();
-              },
-              child: Container(
-                decoration: BoxDecoration(
-                  shape: BoxShape.rectangle,
-                  borderRadius: BorderRadius.circular(10),
-                  color: Colors.green
-                ),
-                padding: const EdgeInsets.only(left: 18,right: 18,top: 10,bottom: 10),
-
-                child: const Text("Next",style: TextStyle(fontSize: 18,color: Colors.white),),
-              ),
-            )
+            // InkWell(
+            //   onTap: (){
+            //     setState(() {
+            //
+            //     });
+            //    // widget.controller.next();
+            //   },
+            //   child: Container(
+            //     decoration: BoxDecoration(
+            //       shape: BoxShape.rectangle,
+            //       borderRadius: BorderRadius.circular(10),
+            //       color: Colors.green
+            //     ),
+            //     padding: const EdgeInsets.only(left: 18,right: 18,top: 10,bottom: 10),
+            //
+            //     child: const Text("Next",style: TextStyle(fontSize: 18,color: Colors.white),),
+            //   ),
+            // )
           ],
         ),
       ),