+++ lib/src/button/stepper_button.dart
@@ -3,6 +3,8 @@
 ///
 ///
 ///
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:stepper_a/src/button/stepper_a_button_style.dart';
 
@@ -36,7 +38,9 @@ class StepperAButtonWidget extends StatelessWidget {
         onTap();
       },
       child: Container(
-        margin: const EdgeInsets.only(bottom: 12, left: 12, right: 12),
+        margin: Platform.isAndroid
+            ?const EdgeInsets.only(bottom: 12, left: 12, right: 12)
+            :const EdgeInsets.only(bottom: 20, left: 12, right: 12),
         width: _stepperAButton.width,
         height: _stepperAButton.height,
         alignment: Alignment.center,